+++ Spryker/Sniffs/Commenting/DocBlockParamNotJustNullSniff.php
@@ -1,0 +1,90 @@
+<?php
+
+namespace Spryker\Sniffs\Commenting;
+
+use PHP_CodeSniffer_File;
+use Spryker\Sniffs\AbstractSniffs\AbstractSprykerSniff;
+use Spryker\Tools\Traits\CommentingTrait;
+use Spryker\Tools\Traits\SignatureTrait;
+
+/**
+ * Makes sure doc block param types are never just `null`, but always another type and optionally nullable on top.
+ *
+ * @author Mark Scherer
+ * @license MIT
+ */
+class DocBlockParamNotJustNullSniff extends AbstractSprykerSniff
+{
+
+    use CommentingTrait;
+    use SignatureTrait;
+
+    /**
+     * @inheritDoc
+     */
+    public function register()
+    {
+        return [
+            T_FUNCTION,
+        ];
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function process(PHP_CodeSniffer_File $phpCsFile, $stackPointer)
+    {
+        $tokens = $phpCsFile->getTokens();
+
+        $docBlockEndIndex = $this->findRelatedDocBlock($phpCsFile, $stackPointer);
+
+        if (!$docBlockEndIndex) {
+            return;
+        }
+
+        $methodSignature = $this->getMethodSignature($phpCsFile, $stackPointer);
+        if (!$methodSignature) {
+            return;
+        }
+
+        $docBlockStartIndex = $tokens[$docBlockEndIndex]['comment_opener'];
+
+        $paramCount = 0;
+        for ($i = $docBlockStartIndex + 1; $i < $docBlockEndIndex; $i++) {
+            if ($tokens[$i]['type'] !== 'T_DOC_COMMENT_TAG') {
+                continue;
+            }
+            if (!in_array($tokens[$i]['content'], ['@param'])) {
+                continue;
+            }
+
+            if (empty($methodSignature[$paramCount])) {
+                continue;
+            }
+            $methodSignatureValue = $methodSignature[$paramCount];
+            $paramCount++;
+
+            $classNameIndex = $i + 2;
+
+            if ($tokens[$classNameIndex]['type'] !== 'T_DOC_COMMENT_STRING') {
+                // Let another sniffer take care of the missing type
+                continue;
+            }
+
+            $content = $tokens[$classNameIndex]['content'];
+
+            $appendix = '';
+            $spaceIndex = strpos($content, ' ');
+            if ($spaceIndex) {
+                $appendix = substr($content, $spaceIndex);
+                $content = substr($content, 0, $spaceIndex);
+            }
+            if (empty($content) || $content !== 'null') {
+                continue;
+            }
+
+            $phpCsFile->addError('"null" as only param type does not make sense', $classNameIndex, 'NotJustNull');
+        }
+    }
+
+}
